made DTO one class per file for best practices and imported lombok for getters and setters
