+++ src/main/java/com/exercisenow/enterprise/dto/Workout.java
@@ -1,0 +1,24 @@
+package com.exercisenow.enterprise.dto;
+
+import lombok.Data;
+
+import java.util.Date;
+
+public @Data
+class Workout {
+    private int workoutID;
+    private int userID;
+    private String type;
+    private double duration; // in minutes
+    private String intensity;
+    private double caloriesBurned;
+    private Date date;
+
+    public void logWorkout() {
+        // Implementation for logging workout
+    }
+
+    public void calculateCalories() {
+        // Implementation for calculating calories burned
+    }
+}
